requirements.txt
asgiref==3.5.2
boto==2.49.0
boto3==1.24.92
botocore==1.27.92
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-paypal==2.0
django-storages==1.13.1
idna==3.4
jmespath==1.0.1
paypal==1.2.5
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
urllib3==1.26.12
